Se implementa herencia en los fireballs
Se generan dos tipos de fireballs que heredan de una clase madre común
