--- PVJ2-proyecto2D/Assets/Prefabs/Enemigos/Fireball.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 56c4b9794b787b04c8bf622c782b6163
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
