+++ PVJ2-proyecto2D/Assets/Prefabs/Enemigos/FireballOblicuo.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 344c81922f05dba42b30a4d43e95e879
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
