--- PVJ2-proyecto2D/Assets/Scripts/Enemigos/DisparoFireball.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 13d8ff1651b0775439f6da91df11a298
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
