+++ PVJ2-proyecto2D/Assets/Scripts/Enemigos/Fireballs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d46da67f2844c95439b45c0b36122ca1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
