+++ PVJ2-proyecto2D/Assets/Scripts/Enemigos/Fireballs/DisparoOblicuo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7b4e59e138f7254e8762931fa513b4a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
