+++ PVJ2-proyecto2D/Assets/Scripts/Enemigos/Fireballs/DisparoRectoFireball.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13d8ff1651b0775439f6da91df11a298
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
