+++ PVJ2-proyecto2D/Assets/Scripts/Enemigos/Fireballs/Fireball.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9277035a4acccca43a05842b1932ccca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
