+++ PVJ2-proyecto2D/Assets/Scripts/Enemigos/Fireballs/GeneradorObjetoLoopWithPool.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12f59666c394660488a651662b248b0a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
