--- PVJ2-proyecto2D/Assets/Scripts/Enemigos/GeneradorObjetoAleatorio.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7e33205e617138e47898560122865851
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
