--- PVJ2-proyecto2D/Assets/Scripts/Enemigos/GeneradorObjetoLoopWithPool.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 12f59666c394660488a651662b248b0a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
