+++ PVJ2-proyecto2D/Assets/Scripts/Enemigos/Jefe Final.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 72cc48dffeb94824da6f882368d591de
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
