+++ PVJ2-proyecto2D/Assets/Scripts/Enemigos/Jefe Final/DisparoBola.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e37ff258717ff5d4fb0f734bb6606195
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
