+++ PVJ2-proyecto2D/Assets/Scripts/Enemigos/Jefe Final/JefeFinal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a41ec92b5eec9f94fb40d7b13f3073fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
