--- PVJ2-proyecto2D/Assets/Scripts/Enemigos/ObjectPool.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a83e35623439ffa4da3d233fcd03062e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
